# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subway Guatemala Customer API",
        "description": "API REST para la aplicación móvil de clientes de Subway Guatemala. Incluye autenticación multi-canal (email/password, Google OAuth, Apple Sign-In), gestión de perfil, dispositivos FCM y futuras funcionalidades de pedidos y lealtad.",
        "contact": {
            "email": "dev@subwayguatemala.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Application Server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/oauth/google": {
            "post": {
                "tags": [
                    "OAuth"
                ],
                "summary": "Login/Register with Google",
                "description": "Authenticates or registers customer using Google OAuth id_token. Mobile app obtains id_token from Google Sign-In SDK, backend verifies with Google and returns Sanctum token. Creates new account if email doesn't exist.",
                "operationId": "7de0059c713c7fba77ef30a111c1522e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_token"
                                ],
                                "properties": {
                                    "id_token": {
                                        "description": "Google OAuth id_token obtained from Google Sign-In SDK",
                                        "type": "string",
                                        "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjhlY..."
                                    },
                                    "device_name": {
                                        "description": "Device name for token",
                                        "type": "string",
                                        "example": "Pixel 7 Pro",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso."
                                        },
                                        "data": {
                                            "properties": {
                                                "customer": {
                                                    "$ref": "#/components/schemas/Customer"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123xyz..."
                                                },
                                                "is_new_customer": {
                                                    "description": "True if account was just created",
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid token or account conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token de Google inválido o expirado."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/v1/auth/oauth/apple": {
            "post": {
                "tags": [
                    "OAuth"
                ],
                "summary": "Login/Register with Apple",
                "description": "Authenticates or registers customer using Apple Sign-In id_token. Mobile app obtains id_token from Apple Sign-In SDK, backend verifies with Apple and returns Sanctum token. Creates new account if email doesn't exist.",
                "operationId": "df8a9b89f5e6a009a05c437fdd4c39d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_token"
                                ],
                                "properties": {
                                    "id_token": {
                                        "description": "Apple OAuth id_token obtained from Apple Sign-In SDK",
                                        "type": "string",
                                        "example": "eyJraWQiOiJlWGF1bm1MIiwiYWxnIj..."
                                    },
                                    "device_name": {
                                        "description": "Device name for token",
                                        "type": "string",
                                        "example": "iPhone 14 Pro",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuenta creada exitosamente."
                                        },
                                        "data": {
                                            "properties": {
                                                "customer": {
                                                    "$ref": "#/components/schemas/Customer"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123xyz..."
                                                },
                                                "is_new_customer": {
                                                    "description": "True if account was just created",
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid token or account conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token de Apple inválido o expirado."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Customer": {
                "title": "Customer",
                "description": "Customer model with profile information",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+50212345678",
                        "nullable": true
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "oauth_provider": {
                        "type": "string",
                        "enum": [
                            "local",
                            "google",
                            "apple"
                        ],
                        "example": "local"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z",
                        "nullable": true
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T15:45:00Z",
                        "nullable": true
                    },
                    "customer_type_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-10T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T15:45:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtain from /api/v1/auth/login, /api/v1/auth/register, or OAuth endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}